Start work on integrating nops in
This has been made very complicated by the poor logic for handling
inline chunks :) - because somehow, it never got tested before now.

So after fixing up the inline chunks logic, it moooostly works. The main
problem to solve is that the patching mechanisms only realllly work with
function blocks (for whatever reason that is), so we need to resolve all
inline blocks before starting tracing and patch derivation.

This in turn causes some fun problems - in doing this, it's very easy to
end up in bizarre states where a function call node points to a block
that doesn't really exist anymore, because we've already patched it!
Short of reworking the entire way that blocks work, I've added a hacky
repair_call method to fixup the call graph to what the programmer (me)
expects.